Allow tensorflow dataset to fetch chunks
If using the tensorflow dataset on a hub dataset which consist of
multiple samples with small size, the iteration becomes considerably
faster if using fetch_chunks=True when getting the data. This commit
adds the ability to pass that option when creating the tensorflow
dataset from the hub dataset.

On an MNIST dataset the result is 25 times faster iteration.